backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
groq==0.4.1
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6